+++ classes/WpMatomo/Referral.php
@@ -46,10 +46,10 @@ class Referral {
 		$self = $this;
 
 		add_action( 'wp_ajax_matomo_referral_dismiss_admin_notice', function () use ($self) {
-			if ($self->should_show() && $self->can_refer()) {
+			if (is_admin() && $self->should_show() && $self->can_refer()) {
 				// no need for an nonce check here as it's nothing critical
 				if (!empty($_POST['forever'])) {
-					$self->never_show_again();
+					$self->dismiss_forever();
 				} else {
 					$self->dismiss();
 				}
@@ -57,12 +57,19 @@ class Referral {
 		});
 		add_action('admin_notices', function () use ($self) {
 			if ($self->can_refer() && $self->should_show_on_screen()) {
-				include 'views/referral.php';
+				$self->render();
 			}
 		});
 	}
 
+	public function render() {
+        include 'views/referral.php';
+    }
+
 	public function should_show_on_screen() {
+	    if (!is_admin()) {
+	        return false;
+        }
 		$screen = get_current_screen();
 		return $screen && $screen->id && strpos($screen->id, 'matomo-') === 0;
 	}
@@ -71,7 +78,7 @@ class Referral {
 		return current_user_can(Capabilities::KEY_VIEW);
 	}
 
-	public function never_show_again() {
+	public function dismiss_forever() {
 		$tenYears = 60 * 60 * 24 * 365 * 10;
 		update_option(self::OPTION_NAME_REFERRAL_DISMISSED, $this->time + $tenYears);
 	}
@@ -80,16 +87,26 @@ class Referral {
 		update_option(self::OPTION_NAME_REFERRAL_DISMISSED, $this->time, true);
 	}
 
+	public function get_last_dismissed() {
+	    return get_option(self::OPTION_NAME_REFERRAL_DISMISSED);
+    }
+
+    private function get_days_in_seconds($num_days)
+    {
+        return 60 * 60 * 24 * $num_days;
+    }
+
 	public function should_show() {
-		$dismissed = get_option(self::OPTION_NAME_REFERRAL_DISMISSED);
+		$dismissed = $this->get_last_dismissed();
 
 		if (!$dismissed) {
-			// the first time we check... basically after install... should not be executed for another 90 days
+			// the first time we check... we set it back 30 days cause we want to see first rating after 60 days
+			$this->time = $this->time - $this->get_days_in_seconds(30);
 			$this->dismiss();
 			return false;
 		}
 
-		$ninetyDaysInSeconds = 60 * 60 * 24 * 90;
+		$ninetyDaysInSeconds = $this->get_days_in_seconds(90);
 
 		if ($this->time > ($dismissed + $ninetyDaysInSeconds)) {
 			return true;